+++ CotEditor/Sources/Document Window/Text View/EditorTextView+Indenting.swift
@@ -8,7 +8,7 @@
 //
 //  ---------------------------------------------------------------------------
 //
-//  © 2014-2024 1024jp
+//  © 2014-2025 1024jp
 //
 //  Licensed under the Apache License, Version 2.0 (the "License");
 //  you may not use this file except in compliance with the License.
@@ -129,12 +129,14 @@ extension Indenting {
     
     
     /// Standardizes indentation of given ranges in the selected ranges.
+    ///
+    /// - Parameter style: The indentation style to change.
     func convertIndentation(style: IndentStyle) {
         
         guard
             self.tabWidth > 0,
             let selectedRanges = self.rangesForUserTextChange?.map(\.rangeValue),
-            let textEditing = self.string.convertIndentation(to: self.indentStyle, indentWidth: self.tabWidth, in: selectedRanges)
+            let textEditing = self.string.convertIndentation(to: style, indentWidth: self.tabWidth, in: selectedRanges)
         else { return }
         
         self.edit(with: textEditing, actionName: String(localized: "Convert Indentation", table: "MainMenu"))